+++ utils.go
@@ -15,7 +15,7 @@ func IsError(r *http.Response) bool {
 	return r.StatusCode >= 400
 }
 
-func intToTime(i int) time.Time {
+func intToTime(i int64) time.Time {
 	return time.Unix(i, 0)
 }
